+++ CHANGELOG.md
@@ -714,9 +714,9 @@ for the initial code allowing to compare two Griffe trees.
 
 ### Breaking changes
 
-- All parameters of the [`load_git`][griffe.git.load_git] function, except `module`, are now keyword-only.
-- Parameter `try_relative_path` of the [`load_git`][griffe.git.load_git] function was removed.
-- Parameter `commit` was renamed `ref` in the [`load_git`][griffe.git.load_git] function.
+- All parameters of the [`load_git`][griffe.loader.load_git] function, except `module`, are now keyword-only.
+- Parameter `try_relative_path` of the [`load_git`][griffe.loader.load_git] function was removed.
+- Parameter `commit` was renamed `ref` in the [`load_git`][griffe.loader.load_git] function.
 - Parameter `commit` was renamed `ref` in the `tmp_worktree` helper, which will probably become private later.
 - Parameters `ref` and `repo` switched positions in the `tmp_worktree` helper.
 - All parameters of the [`resolve_aliases`][griffe.loader.GriffeLoader.resolve_aliases] method are now keyword-only.